Fix bugs and re-organize code
