+++ optimization_specialist_train.py
@@ -3,26 +3,25 @@ from evoman.environment import Environment
 import time
 import os
 import numpy as np
-from operators.selection import evaluate, parent_selection, survivor_selection
-from operators.variation import crossover, mutate
+from operators import crossover, mutate, initialize_population, evaluate, survivor_selection
 
 # Define a set of constants
 POPULATION_SIZE = 100
-INDIVIDUAL_SIZE = 265
+INDIVIDUAL_SIZE = 5
 TOTAL_GENERATIONS = 50
 
 # Set Parameters
 lower_bound = -1
 upper_bound = 1
-k = 5  # tournament_size
+tournament_size = 5
 mutation_rate = 0.2
 mutation_sigma = 0.1
 selection_pressure = 1.5
 
-# initialize simulation
+# Initialize simulation
 env = Environment()
 
-# check environment state
+# Log the environment state
 env.state_to_log()
 
 # set time marker
@@ -32,16 +31,18 @@ ini = time.time()
 generation_count = 0
 
 # Initialize population loading old solutions or generating new ones
-experiment_name = ''
+experiment_name = 'experiments'
 if not os.path.exists(experiment_name + '/evoman_solstate'):
     print('\nNEW EVOLUTION\n')
     # Generate new population here
-    population = np.random.uniform(lower_bound, upper_bound, (POPULATION_SIZE, INDIVIDUAL_SIZE))
+    population = initialize_population(POPULATION_SIZE, INDIVIDUAL_SIZE, lower_bound, upper_bound)
     population_fitness = evaluate(env, population)
+
+    # Save current population and fitness values
     env.update_solutions([population, population_fitness])
 else:
     print('\nCONTINUING EVOLUTION\n')
-    # Load generation from env
+    # Load env state
     env.load_state()
     population = env.solutions[0]
     population_fitness = env.solutions[1]
@@ -51,6 +52,7 @@ else:
     generation_count = int(generation_file.readline())
     generation_file.close()
 
+# Compute max, mean and std for current population
 best_individual_index = np.argmax(population_fitness)
 mean = np.mean(population_fitness)
 std = np.std(population_fitness)
@@ -71,19 +73,21 @@ print('\n GENERATION ' + str(generation_count)
 
 # Evolution
 while generation_count < TOTAL_GENERATIONS:
-    # Select parents
-    parent1, parent2 = parent_selection(population, k, env)
-
     # Create offspring
-    child1, child2 = crossover(parent1, parent2)
+    tournament_count = int(POPULATION_SIZE / 2)
+    offspring = crossover(env, population, tournament_count, tournament_size)
 
     # Mutate offspring
-    child1 = mutate(child1, rate=mutation_rate, sigma=mutation_sigma)
-    child2 = mutate(child2, rate=mutation_rate, sigma=mutation_sigma)
+    for i in range(len(offspring)):
+        offspring[i] = mutate(offspring[i], rate=mutation_rate, sigma=mutation_sigma)
 
     # Survivor selection
-    population, population_fitness = survivor_selection(population, [child1, child2], population_fitness,
-                                                        selection_pressure)
+    population, population_fitness = survivor_selection(
+        population,
+        offspring,
+        population_fitness,
+        env,
+        selection_pressure)
 
     best_individual_index = np.argmax(population_fitness)
     mean = np.mean(population_fitness)
